+++ piecrust-uplink/src/abi_host/externs.rs
@@ -30,16 +30,14 @@ unsafe extern "C" {
     ) -> i32;
 
     #[link_name = "__piecrust_b_host_query"]
-    pub fn hq_v2(
+    pub fn host_query(
         name: *const u8,
         name_len: u32,
         arg: *const u8,
         arg_len: u32,
     ) -> u32;
-    #[link_name = "__piecrust_b_host_data"]
-    pub fn hd_v2(name: *const u8, name_len: u32) -> u32;
     #[link_name = "__piecrust_b_call"]
-    pub fn c_v2(
+    pub fn call(
         contract_id: *const u8,
         fn_name: *const u8,
         fn_name_len: u32,
@@ -48,33 +46,4 @@ unsafe extern "C" {
         gas_limit: u64,
     ) -> i32;
 
-    #[link_name = "__piecrust_b_emit"]
-    pub fn emit_v2(
-        topic: *const u8,
-        topic_len: u32,
-        data: *const u8,
-        data_len: u32,
-    );
-    #[link_name = "__piecrust_b_feed"]
-    pub fn feed_v2(data: *const u8, data_len: u32);
-
-    #[link_name = "__piecrust_b_caller"]
-    pub fn caller_v2() -> i32;
-    #[link_name = "__piecrust_b_callstack"]
-    pub fn callstack_v2() -> i32;
-    #[link_name = "__piecrust_b_owner"]
-    pub fn owner_v2(contract_id: *const u8) -> i32;
-    #[link_name = "__piecrust_b_self_id"]
-    pub fn self_id_v2();
-
-    pub fn limit() -> u64;
-    pub fn spent() -> u64;
-
-    #[cfg(any(target_arch = "wasm32", target_arch = "wasm64"))]
-    #[link_name = "__piecrust_b_panic"]
-    pub fn panic_v2(message: *const u8, message_len: u32) -> !;
-
-    #[cfg(feature = "debug")]
-    #[link_name = "__piecrust_b_debug"]
-    pub fn hdebug_v2(message: *const u8, message_len: u32);
 }